chore: Move ApplicationClient (http) to new folder/namespace
